# swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "Region": {
        "enum": [
          "East",
          "NorthCentral",
          "NorthWest",
          "SouthCentral",
          "SouthWest"
        ],
        "type": "string"
      },
      "Organization": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "type": "array"
          },
          "fundingSpaces": {
            "items": {
              "$ref": "#/components/schemas/FundingSpace"
            },
            "type": "array"
          }
        },
        "required": ["id", "name"],
        "type": "object",
        "additionalProperties": false
      },
      "Gender": {
        "enum": ["Male", "Female", "Nonbinary", "Unknown", "Unspecified"],
        "type": "string"
      },
      "Child": {
        "properties": {
          "id": {
            "type": "string"
          },
          "sasid": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "suffix": {
            "type": "string"
          },
          "birthdate": {
            "type": "string",
            "format": "date-time"
          },
          "birthTown": {
            "type": "string"
          },
          "birthState": {
            "type": "string"
          },
          "birthCertificateId": {
            "type": "string"
          },
          "americanIndianOrAlaskaNative": {
            "type": "boolean"
          },
          "asian": {
            "type": "boolean"
          },
          "blackOrAfricanAmerican": {
            "type": "boolean"
          },
          "nativeHawaiianOrPacificIslander": {
            "type": "boolean"
          },
          "white": {
            "type": "boolean"
          },
          "hispanicOrLatinxEthnicity": {
            "type": "boolean"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "foster": {
            "type": "boolean"
          },
          "familyId": {
            "type": "number",
            "format": "double"
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "enrollments": {
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "type": "array"
          },
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "c4KFamilyCaseNumber": {
            "type": "number",
            "format": "double"
          },
          "c4KCertificates": {
            "items": {
              "$ref": "#/components/schemas/C4KCertificate"
            },
            "type": "array"
          },
          "authorId": {
            "type": "number",
            "format": "double"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "firstName", "lastName", "organizationId"],
        "type": "object",
        "additionalProperties": false
      },
      "Family": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/Child"
            },
            "type": "array"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "homelessness": {
            "type": "boolean"
          },
          "determinations": {
            "items": {
              "$ref": "#/components/schemas/FamilyDetermination"
            },
            "type": "array"
          },
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "authorId": {
            "type": "number",
            "format": "double"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "organizationId"],
        "type": "object",
        "additionalProperties": false
      },
      "User": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "wingedKeysId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "suffix": {
            "type": "string"
          },
          "orgPermissions": {
            "items": {
              "$ref": "#/components/schemas/OrganizationPermission"
            },
            "type": "array"
          },
          "sitePermissions": {
            "items": {
              "$ref": "#/components/schemas/SitePermission"
            },
            "type": "array"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "type": "array"
          }
        },
        "required": ["id", "wingedKeysId", "firstName", "lastName"],
        "type": "object",
        "additionalProperties": false
      },
      "FamilyDetermination": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "numberOfPeople": {
            "type": "number",
            "format": "double"
          },
          "income": {
            "type": "number",
            "format": "double"
          },
          "determinationDate": {
            "type": "string",
            "format": "date-time"
          },
          "familyId": {
            "type": "number",
            "format": "double"
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "authorId": {
            "type": "number",
            "format": "double"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "familyId"],
        "type": "object",
        "additionalProperties": false
      },
      "Enrollment": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "childId": {
            "type": "string"
          },
          "child": {
            "$ref": "#/components/schemas/Child"
          },
          "siteId": {
            "type": "number",
            "format": "double"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "ageGroup": {
            "$ref": "#/components/schemas/Age"
          },
          "entry": {
            "type": "string",
            "format": "date-time"
          },
          "exit": {
            "type": "string",
            "format": "date-time"
          },
          "exitReason": {
            "type": "string"
          },
          "fundings": {
            "items": {
              "$ref": "#/components/schemas/Funding"
            },
            "type": "array"
          },
          "pastEnrollments": {
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "type": "array"
          },
          "authorId": {
            "type": "number",
            "format": "double"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "childId", "siteId"],
        "type": "object",
        "additionalProperties": false
      },
      "C4KCertificate": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "childId": {
            "type": "string"
          },
          "child": {
            "$ref": "#/components/schemas/Child"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "childId"],
        "type": "object",
        "additionalProperties": false
      },
      "Site": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "titleI": {
            "type": "boolean"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "facilityCode": {
            "type": "number",
            "format": "double"
          },
          "licenseNumber": {
            "type": "number",
            "format": "double"
          },
          "naeycId": {
            "type": "number",
            "format": "double"
          },
          "registryId": {
            "type": "number",
            "format": "double"
          },
          "enrollments": {
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "type": "array"
          }
        },
        "required": ["id", "name", "titleI", "region", "organizationId"],
        "type": "object",
        "additionalProperties": false
      },
      "Age": {
        "enum": ["InfantToddler", "Preschool", "SchoolAge"],
        "type": "string"
      },
      "FundingSource": {
        "enum": ["CDC"],
        "type": "string"
      },
      "Funding": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "enrollmentId": {
            "type": "number",
            "format": "double"
          },
          "enrollment": {
            "$ref": "#/components/schemas/Enrollment"
          },
          "fundingSpace": {
            "$ref": "#/components/schemas/FundingSpace"
          },
          "fundingSpaceId": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "firstReportingPeriodId": {
            "type": "number",
            "format": "double"
          },
          "firstReportingPeriod": {
            "$ref": "#/components/schemas/ReportingPeriod"
          },
          "lastReportingPeriodId": {
            "type": "number",
            "format": "double"
          },
          "lastReportingPeriod": {
            "$ref": "#/components/schemas/ReportingPeriod"
          },
          "authorId": {
            "type": "number",
            "format": "double"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "enrollmentId", "fundingSpaceId"],
        "type": "object",
        "additionalProperties": false
      },
      "FundingTime": {
        "enum": ["Full", "Part", "Split"],
        "type": "string"
      },
      "FundingSpace": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "capacity": {
            "type": "number",
            "format": "double"
          },
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "source": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "ageGroup": {
            "$ref": "#/components/schemas/Age"
          },
          "fundings": {
            "items": {
              "$ref": "#/components/schemas/Funding"
            },
            "type": "array"
          },
          "time": {
            "$ref": "#/components/schemas/FundingTime"
          },
          "timeSplit": {
            "$ref": "#/components/schemas/FundingTimeSplit"
          },
          "timeSplitUtilizations": {
            "items": {
              "$ref": "#/components/schemas/FundingTimeSplitUtilization"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "capacity",
          "organizationId",
          "source",
          "ageGroup",
          "time"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "FundingTimeSplit": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "fundingSpaceId": {
            "type": "number",
            "format": "double"
          },
          "fundingSpace": {
            "$ref": "#/components/schemas/FundingSpace"
          },
          "fullTimeWeeks": {
            "type": "number",
            "format": "double"
          },
          "partTimeWeeks": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["id", "fundingSpaceId", "fullTimeWeeks", "partTimeWeeks"],
        "type": "object",
        "additionalProperties": false
      },
      "ReportingPeriod": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "period": {
            "type": "string",
            "format": "date-time"
          },
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "period",
          "periodStart",
          "periodEnd",
          "dueAt"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "FundingTimeSplitUtilization": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "reportingPeriodId": {
            "type": "number",
            "format": "double"
          },
          "reportingPeriod": {
            "$ref": "#/components/schemas/ReportingPeriod"
          },
          "reportId": {
            "type": "number",
            "format": "double"
          },
          "report": {
            "$ref": "#/components/schemas/CdcReport"
          },
          "fundingSpaceId": {
            "type": "number",
            "format": "double"
          },
          "fundingSpace": {
            "$ref": "#/components/schemas/FundingSpace"
          },
          "fullTimeWeeksUsed": {
            "type": "number",
            "format": "double"
          },
          "partTimeWeeksUsed": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id",
          "reportingPeriodId",
          "reportId",
          "fundingSpaceId",
          "fullTimeWeeksUsed",
          "partTimeWeeksUsed"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CdcReport": {
        "properties": {
          "accredited": {
            "type": "boolean"
          },
          "c4KRevenue": {
            "type": "number",
            "format": "double"
          },
          "retroactiveC4KRevenue": {
            "type": "boolean"
          },
          "familyFeesRevenue": {
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "timeSplitUtilizations": {
            "items": {
              "$ref": "#/components/schemas/FundingTimeSplitUtilization"
            },
            "type": "array"
          },
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "id": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "reportingPeriodId": {
            "type": "number",
            "format": "double"
          },
          "reportingPeriod": {
            "$ref": "#/components/schemas/ReportingPeriod"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "enrollments": {
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "type": "array"
          }
        },
        "required": [
          "accredited",
          "c4KRevenue",
          "familyFeesRevenue",
          "organizationId",
          "id",
          "reportingPeriodId",
          "reportingPeriod"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "OrganizationPermission": {
        "properties": {
          "organizationId": {
            "type": "number",
            "format": "double"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "id": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "number",
            "format": "double"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["organizationId", "id", "userId"],
        "type": "object",
        "additionalProperties": false
      },
      "SitePermission": {
        "properties": {
          "siteId": {
            "type": "number",
            "format": "double"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "id": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "number",
            "format": "double"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["siteId", "id", "userId"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "info": {
    "title": "data-collection",
    "version": "0.0.1",
    "license": {
      "name": "MIT"
    },
    "contact": {}
  },
  "openapi": "3.0.0",
  "paths": {
    "/reports/{reportId}": {
      "get": {
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/reports": {
      "post": {
        "operationId": "Post",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/current": {
      "get": {
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}